# swagger.yaml
openapi: 3.0.0
paths:
...
  "/v1/repos/{id}":
    get:
      operationId: findOneById
      summary: Finds a repo by :id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbRepo"
        "404":
          description: Repository not found
      tags:
        &a8
        - Repository service
  "/v1/repos/{owner}/{repo}":
    get:
      operationId: findOneByOwnerAndRepo
      summary: Finds a repo by :owner and :repo
      parameters:
        - name: owner
          required: true
          in: path
          schema:
            type: string
        - name: repo
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbRepo"
        "404":
          description: Repository not found
      tags: *a8
  /v1/repos/list:
    get:
      operationId: findAll
      summary: Finds all repos and paginates them
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: integer
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 1000
            default: 10
            type: integer
        - name: orderDirection
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/OrderDirectionEnum"
        - name: range
          required: false
          in: query
          description: Range in days
          schema:
            default: 30
            type: integer
        - name: orderBy
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/RepoOrderFieldsEnum"
      responses:
        "200":
          description: Successfully received model list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DbRepo"
      tags: *a8
  /v1/repos/search:
    get:
      operationId: findAllReposWithFilters
      summary: Finds all repos using filters and paginates them
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: integer
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 1000
            default: 10
            type: integer
        - name: orderDirection
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/OrderDirectionEnum"
        - name: range
          required: false
          in: query
          description: Range in days
          schema:
            default: 30
            type: integer
        - name: orderBy
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/RepoOrderFieldsEnum"
        - name: filter
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/InsightFilterFieldsEnum"
        - name: repo
          required: false
          in: query
          schema:
            type: string
        - name: topic
          required: false
          in: query
          schema:
            default: ""
            type: string
        - name: repoIds
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Successfully received model list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DbRepo"
      tags: *a8
...
  /v1/highlights/repos/list:
    get:
      operationId: findAllHighlightRepos
      summary: Finds all highlight repos and paginates them
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: integer
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 1000
            default: 10
            type: integer
        - name: orderDirection
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/OrderDirectionEnum"
        - name: range
          required: false
          in: query
          description: Range in days
          schema:
            default: 30
            type: integer
      responses:
        "200":
          description: Successfully received model list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DbUserHighlightRepo"
      tags: *a24
...
  "/v1/endorsements/repos/{repoOwnerOrUser}":
    get:
      operationId: findAllByRepoOwnerOrUsername
      summary: Finds all endorsements by repo org or username and paginates them
      parameters:
        - name: repoOwnerOrUser
          required: true
          in: path
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: integer
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 1000
            default: 10
            type: integer
        - name: orderDirection
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/OrderDirectionEnum"
        - name: range
          required: false
          in: query
          description: Range in days
          schema:
            default: 30
            type: integer
      responses:
        "200":
          description: Successfully received model list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DbEndorsement"
      tags: *a25
  "/v1/endorsements/repos/{owner}/{repo}":
    get:
      operationId: findAllEndorsementsByRepo
      summary: Finds all endorsements by repo owner or username and paginates them
      parameters:
        - name: owner
          required: true
          in: path
          schema:
            type: string
        - name: repo
          required: true
          in: path
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: integer
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 1000
            default: 10
            type: integer
        - name: orderDirection
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/OrderDirectionEnum"
        - name: range
          required: false
          in: query
          description: Range in days
          schema:
            default: 30
            type: integer
      responses:
        "200":
          description: Successfully received model list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DbEndorsement"
      tags: *a25
...
  "/v1/bake/repos/{id}":
    get:
      operationId: findBakedRepoById
      summary: Finds a baked repo by :id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbBakedRepo"
        "404":
          description: Baked repository not found
      tags: *a27
  /v1/bake/repos/list:
    get:
      operationId: listAllBakedRepos
      summary: Finds all baked repos and paginates them
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: integer
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 1000
            default: 10
            type: integer
        - name: orderDirection
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/OrderDirectionEnum"
        - name: range
          required: false
          in: query
          description: Range in days
          schema:
            default: 30
            type: integer
      responses:
        "200":
          description: Successfully received model list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DbBakedRepo"
      tags: *a27
...
  "/v1/bake/commits/list/repo/{id}":
    get:
      operationId: listAllCommitsByBakedRepoId
      summary: Finds all commits by baked repo :id and paginates them
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: integer
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 1000
            default: 10
            type: integer
        - name: orderDirection
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/OrderDirectionEnum"
        - name: range
          required: false
          in: query
          description: Range in days
          schema:
            default: 30
            type: integer
        - name: id
          required: true
          in: path
          schema:
            type: integer
      responses:
        "200":
          description: Successfully received model list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DbCommits"
      tags: *a27
...
info:
  title: "@open-sauced/api.opensauced.pizza"
  description: >-
    
    ## Swagger-UI API Documentation


    This REST API can be used to create, read, update or delete data from the Open Sauced community platform.

    The Swagger-UI provides useful information to get started and an overview of all available resources.

    Each API route is clickable and has their own detailed description on how to use it.

    The base URL for the API is [api.opensauced.pizza](https://api.opensauced.pizza).


    [comment]: # (TODO: add bearer auth information)


    ## Rate limiting


    Every IP address is allowed to perform 5000 requests per hour.

    This is measured by saving the date of the initial request and counting all requests in the next hour.

    When an IP address goes over the limit, HTTP status code 429 is returned.

    The returned HTTP headers of any API request show the current rate limit status:


    header | description

    --- | ---

    `X-RateLimit-Limit` | The maximum number of requests allowed per hour

    `X-RateLimit-Remaining` | The number of requests remaining in the current rate limit window

    `X-RateLimit-Reset` | The date and time at which the current rate limit window resets in [UTC epoch seconds](https://en.wikipedia.org/wiki/Unix_time)


    [comment]: # (TODO: add pagination information)


    ## Common response codes


...

    The table below shows most common response codes you can receive from our endpoints.


...

...

...

...

...

...

...

...

...

...


...
servers:
...
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    PageMetaDto:
      type: object
      properties:
        page:
          type: integer
          description: The current page
          example: 1
        limit:
          type: integer
          description: The number of items per page
          example: 10
        itemCount:
          type: integer
          description: The number of items in the collection
          example: 100
        pageCount:
          type: integer
          description: The number of pages in the collection
          example: 10
        hasPreviousPage:
          type: boolean
          description: Flag indicating if there is a previous page
          example: false
        hasNextPage:
          type: boolean
          description: Flag indicating if there is a next page
          example: true
      required:
        - page
        - limit
        - itemCount
        - pageCount
        - hasPreviousPage
        - hasNextPage
    PageDto:
      type: object
      properties:
        data:
          type: array
          items:
            type: array
        meta:
          $ref: "#/components/schemas/PageMetaDto"
      required:
        - data
        - meta
...
    OrderDirectionEnum:
      type: string
      enum:
        - ASC
        - DESC
...
    DbRepo:
      type: object
      properties:
        id:
          type: integer
          description: Repository identifier
          example: 71359796
        user_id:
          type: integer
          description: Owner user identifier
          example: 57568598
        size:
          type: integer
          description: Total size in bytes
          example: 274322
        issues:
          type: integer
          description: Total number of issues
          example: 274
        stars:
          type: integer
          description: Total number of stars
          example: 5
        forks:
          type: integer
          description: Total number of forks
          example: 1
        watchers:
          type: integer
          description: Total number of watchers
          example: 5473
        subscribers:
          type: integer
          description: Total number of subscribers
          example: 11756
        network:
          type: integer
          description: Total number of network usages
          example: 4
        is_fork:
          type: boolean
          description: Flag indicating repo is a fork
          example: false
        is_private:
          type: boolean
          description: Flag indicating repo is private
          example: false
        is_template:
          type: boolean
          description: Flag indicating repo is a template
          example: false
        is_archived:
          type: boolean
          description: Flag indicating repo is archived
          example: false
        is_disabled:
          type: boolean
          description: Flag indicating repo is disabled
          example: false
        has_issues:
          type: boolean
          description: Flag indicating repo has issues enabled
          example: false
        has_projects:
          type: boolean
          description: Flag indicating repo has projects enabled
          example: false
        has_downloads:
          type: boolean
          description: Flag indicating repo has downloads enabled
          example: false
        has_wiki:
          type: boolean
          description: Flag indicating repo has wiki enabled
          example: false
        has_pages:
          type: boolean
          description: Flag indicating repo has pages enabled
          example: false
        has_discussions:
          type: boolean
          description: Flag indicating repo has discussions enabled
          example: false
        created_at:
          format: date-time
          type: string
          description: Timestamp representing repository creation
          example: 2016-10-19 13:24:51.000000
        updated_at:
          format: date-time
          type: string
          description: Timestamp representing repository last update
          example: 2022-08-28 22:04:29.000000
        pushed_at:
          format: date-time
          type: string
          description: Timestamp representing repository last push
          example: 2022-08-28 22:04:39.000000
        default_branch:
          type: string
          description: Repository default branch
          example: main
        node_id:
          type: string
          description: Repository GQL node id
          example: MDEwOlJlcG9zaXRvcnk3MTM1OTc5Ng==
        git_url:
          type: string
          description: Repository git url
          example: git://github.com/open-sauced/insights.git
        ssh_url:
          type: string
          description: Repository ssh url
          example: git@github.com:open-sauced/insights.git
        clone_url:
          type: string
          description: Repository clone url
          example: https://github.com/open-sauced/insights.git
        svn_url:
          type: string
          description: Repository svn url
          example: https://github.com/open-sauced/insights
        mirror_url:
          type: string
          description: Repository mirror url
          example: null
        name:
          type: string
          description: Repository unique name
          example: open-sauced
        full_name:
          type: string
          description: Repository full name
          example: open-sauced/insights
        description:
          type: string
          description: Repository short description
          example: 🍕This is a project to identify your next open source contribution! 🍕
        language:
          type: string
          description: Repository programming language
          example: JavaScript
        license:
          type: string
          description: Repository SPDX license
          example: MIT
        url:
          type: string
          description: Repository GitHub linked URL
          example: https://api.github.com/repos/open-sauced/insights
        homepage:
          type: string
          description: Repository GitHub homepage
          example: https://app.opensauced.pizza
        topics:
          description: Repository GitHub topics
          example:
            - open-sauced
            - open-source
            - github
          type: array
          items:
            type: string
        open_prs_count:
          type: integer
          description: Repository number of open PRs
          example: 5
        closed_prs_count:
          type: integer
          description: Repository number of closed PRs
          example: 173
        merged_prs_count:
          type: integer
          description: Repository number of merged PRs
          example: 276
        draft_prs_count:
          type: integer
          description: Repository number of draft PRs
          example: 2
        spam_prs_count:
          type: integer
          description: Repository number of spam PRs
          example: 2
        pr_velocity_count:
          type: integer
          description: Repository average open/close time for PRs
          example: 0
        pr_active_count:
          type: integer
          description: Number of non-closed PRs updated within the day range
          example: 0
      required:
        - id
        - user_id
        - size
        - issues
        - stars
        - forks
        - watchers
        - subscribers
        - network
        - is_fork
        - is_private
        - is_template
        - is_archived
        - is_disabled
        - has_issues
        - has_projects
        - has_downloads
        - has_wiki
        - has_pages
        - has_discussions
        - default_branch
        - node_id
        - git_url
        - ssh_url
        - clone_url
        - svn_url
        - name
        - full_name
        - description
        - language
        - license
        - url
        - homepage
        - topics
...
    DbUserHighlightRepo:
      type: object
      properties:
        full_name:
          type: string
          description: Highlight Repo Full Name
          example: open-sauced/insights
...
    DbEndorsement:
      type: object
      properties:
        id:
          type: string
          description: Endorsement identifier
          example: uuid-v4
        creator_user_id:
          type: integer
          description: Endorsement Creator User ID
          example: 237133
        recipient_user_id:
          type: integer
          description: Endorsement Recipient User ID
          example: 31333
        repo_id:
          type: integer
          description: Repository ID
          example: 78133
        source_comment_url:
          type: string
          description: Endorsement Source Comment URL
        source_context_url:
          type: string
          description: Endorsement Source Context URL
        type:
          type: string
          description: Endorsement Type
          example: doc
        created_at:
          format: date-time
          type: string
          description: Timestamp representing endorsement creation
          example: 2022-10-19 13:24:51.000000
      required:
        - id
        - repo_id
        - source_context_url
        - type
...
    InsightFilterFieldsEnum:
      type: string
      enum:
        - recent
        - top-100-repos
        - minimum-5-contributors
        - most-active
        - most-spammed
...
    RepoOrderFieldsEnum:
      type: string
      enum:
        - issues
        - stars
        - watchers
        - subscribers
        - created_at
        - updated_at
        - name
        - contributionsCount
        - votesCount
        - submissionsCount
        - stargazersCount
        - starsCount
...
    DbBakedRepo:
      type: object
      properties:
        id:
          type: integer
          description: Baked repository identifier
          example: 71359796
        clone_url:
          type: string
          description: Repository clone url
          example: https://github.com/open-sauced/insights.git
        repo_id:
          type: integer
          description: Repository identifier
          example: 12345678
      required:
        - id
        - clone_url
        - repo_id
...
    DbCommits:
      type: object
      properties:
        id:
          type: integer
          description: Commit identifier
          example: 71359796
        commit_hash:
          type: string
          description: Hash for the commit
          example: 5e7c6c7af42d38c57f363c564c58007da448c443
        commit_date:
          type: string
          description: Date for the commit
          example: 5e7c6c7af42d38c57f363c564c58007da448c443
        baked_repo_id:
          type: integer
          description: Baked repo identifier
          example: 57568598
        commit_author_id:
          type: integer
          description: Commit author identifier
          example: 9876543
      required:
        - id
        - commit_hash
        - commit_date
        - baked_repo_id
        - commit_author_id